modify on change logger format and time format

+++ src/Connector.cpp
@@ -110,7 +110,7 @@ void Connector::uv_read_callback( uv_stream_t * stream , ssize_t nread , const u
     {
         //LOG_DEBUG_UV( nread );
         session->error_.Code( nread );
-        session->error_.Message( uv_strerror(nread) );
+        session->error_.Message( uv_strerror( ( int ) nread ) );
 
         uv_close( ( uv_handle_t* ) &session->uv_tcp_ , Connector::uv_close_callback );
         return;
